+++ .changeset/rename-configs-to-rspack.md
@@ -1,0 +1,5 @@
+---
+"arui-scripts": patch
+---
+
+Внутренние конфиги сборки переименованы `webpack.*.ts` → `rspack.*.ts` (сборка давно работает на Rspack). Обновлены все внутренние импорты. Изменение внутреннее: эти модули не входят в публичный `exports`, публичный API не затронут.
