+++ packages/arui-scripts/src/commands/bundle-analyze/index.ts
@@ -7,7 +7,7 @@ import {
 import { BundleAnalyzerPlugin } from 'webpack-bundle-analyzer';
 
 import { configs } from '../../configs/app-configs';
-import { webpackClientConfig } from '../../configs/webpack.client.prod';
+import { webpackClientConfig } from '../../configs/rspack.client.prod';
 import { makeTmpDir } from '../util/make-tmp-dir';
 
 type BundleAnalyzerStatsOptions = {
