+++ packages/arui-scripts/src/commands/start-prod/client.ts
@@ -1,4 +1,4 @@
-import { webpackClientConfig } from '../../configs/webpack.client.prod';
+import { webpackClientConfig } from '../../configs/rspack.client.prod';
 import { loadBrowserslist } from '../util/load-browserslist';
 import { runClientDevServer } from '../util/run-client-dev-server';
