+++ packages/arui-scripts/src/commands/start/client.ts
@@ -1,4 +1,4 @@
-import { config as clientConfig } from '../../configs/webpack.client.dev';
+import { config as clientConfig } from '../../configs/rspack.client.dev';
 import { loadBrowserslist } from '../util/load-browserslist';
 import { runClientDevServer } from '../util/run-client-dev-server';
