+++ packages/arui-scripts/src/commands/start/server.ts
@@ -1,5 +1,5 @@
 import { supportingNode } from '../../configs/supporting-node';
-import { webpackServerConfig } from '../../configs/webpack.server.dev';
+import { webpackServerConfig } from '../../configs/rspack.server.dev';
 import { runServerWatchCompiler } from '../util/run-server-watch-compiler';
 
 process.env.BROWSERSLIST = supportingNode.join(',');
