+++ packages/arui-scripts/src/configs/rspack.client.ts
