+++ packages/arui-scripts/src/configs/rspack.server.dev.ts
@@ -1,10 +1,10 @@
 import { applyOverrides } from './util/apply-overrides';
 import { findLoader } from './util/find-loader';
 import { createFindPluginFunction } from './util/find-plugin';
-import { createServerConfig } from './webpack.server';
+import { createServerConfig } from './rspack.server';
 
 export const webpackServerConfig = applyOverrides(
-    ['webpack', 'webpackServer', 'webpackProd', 'webpackServerProd'],
-    createServerConfig('prod'),
+    ['webpack', 'webpackServer', 'webpackDev', 'webpackServerDev'],
+    createServerConfig('dev'),
     { findLoader, findPlugin: createFindPluginFunction<'server'>() },
 );
